refactor: getTemporalStatusとstartAtMsをschedule featureに共通化
index.page.tsxとproblem/group.tsで重複していたgetTemporalStatus・
startAtMs関数をfeatures/schedule/temporal.tsに切り出し、
両方から共有ユーティリティとして利用するように変更。
