+++ frontend/packages/contestant/app/features/problem/group.ts
@@ -1,7 +1,9 @@
 import type { Problem, ScheduleEntry } from "@ictsc/proto/contestant/v1";
-import { timestampDate } from "@bufbuild/protobuf/wkt";
-
-export type ScheduleTemporalStatus = "past" | "current" | "future";
+import {
+  getTemporalStatus,
+  startAtMs,
+  type ScheduleTemporalStatus,
+} from "../schedule";
 
 export type GroupScheduleInfo = {
   name: string;
@@ -14,20 +16,6 @@ export type ProblemGroup = {
   problems: Problem[];
   hasSubmittableProblem: boolean;
 };
-
-function getTemporalStatus(
-  entry: ScheduleEntry,
-  now: Date,
-): ScheduleTemporalStatus {
-  if (entry.endAt != null && now >= timestampDate(entry.endAt)) return "past";
-  if (entry.startAt != null && now >= timestampDate(entry.startAt))
-    return "current";
-  return "future";
-}
-
-function startAtMs(entry: ScheduleEntry): number {
-  return entry.startAt != null ? timestampDate(entry.startAt).getTime() : 0;
-}
 
 export function groupProblems(
   problems: Problem[],
